+++ app/components/base/property-list/bool-property/bool-property.html
@@ -1,0 +1,7 @@
+<bl-property-field>
+    <div propertyLabel>{{label}}</div>
+    <bl-property-content>
+        <i class="fa" [class.fa-toggle-on]="value" [class.fa-toggle-off]="!value"></i>
+        {{ text }}
+    </bl-property-content>
+</bl-property-field>
