+++ app/components/base/property-list/bool-property/index.ts
@@ -1,0 +1,1 @@
+export * from "./bool-property.component";
