+++ app/components/base/property-list/link-property/index.ts
@@ -1,0 +1,1 @@
+export * from "./link-property.component";
