+++ app/components/base/property-list/link-property/link-property.component.ts
@@ -1,0 +1,14 @@
+import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
+
+@Component({
+    selector: "bl-link-property",
+    templateUrl: "link-property.html",
+    changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class LinkPropertyComponent {
+    @Input() public label: string;
+
+    @Input() public value: string;
+
+    @Input() public link: string;
+}
