+++ app/components/base/property-list/link-property/link-property.html
@@ -1,0 +1,4 @@
+<bl-property-field>
+    <label>{{label}}</label>
+    <a class="value link" [routerLink]="link" [title]="value">{{value}}</a>
+</bl-property-field>
