+++ app/components/base/property-list/property-field/property-field.component.ts
@@ -1,10 +1,11 @@
-import { Component } from "@angular/core";
+import { ChangeDetectionStrategy, Component } from "@angular/core";
 
 import "./property-field.scss";
 
 @Component({
     selector: "bl-property-field",
     templateUrl: "property-field.html",
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PropertyFieldComponent {
